Move IrqCell into kernel_api
